+++ tests/properties.rs
@@ -13,7 +13,7 @@ enum Token {
 }
 
 rustlex! PropertiesLexer {
-    property depth:int = 0;
+    property depth:isize = 0;
     let OPEN = '(';
     let CLOSE = ')';
     OPEN => |lexer:&mut PropertiesLexer<R>| { lexer.depth += 1; Some(Open) }
